+++ src/process/serveData.go
@@ -24,7 +24,7 @@ import (
 )
 
 // Serve serves data on a per process basis
-func Serve(process *Process, dataChannel chan *Process, endChannel chan os.Signal, refreshRate int32, wg *sync.WaitGroup) {
+func Serve(process *Process, dataChannel chan *Process, endChannel chan os.Signal, refreshRate uint64, wg *sync.WaitGroup) {
 	for {
 		select {
 		case <-endChannel:
@@ -40,7 +40,7 @@ func Serve(process *Process, dataChannel chan *Process, endChannel chan os.Signa
 
 }
 
-func ServeProcs(dataChannel chan []*proc.Process, endChannel chan os.Signal, refreshRate int32, wg *sync.WaitGroup) {
+func ServeProcs(dataChannel chan []*proc.Process, endChannel chan os.Signal, refreshRate uint64, wg *sync.WaitGroup) {
 	for {
 		select {
 		case <-endChannel:
